deploy_contract.py
from web3 import Web3
import os
from dotenv import load_dotenv

load_dotenv()

def deploy():
    # Connect to BSC
    rpc_url = os.getenv("BSC_RPC_URL", "https://bsc-dataseed.binance.org/")
    w3 = Web3(Web3.HTTPProvider(rpc_url))
    
    if not w3.is_connected():
        print("Failed to connect to BSC")
        return

    print(f"Connected to BSC: {w3.client_version}")
    
    # NOTE: This is a template. Real deployment requires a private key.
    # private_key = os.getenv("PRIVATE_KEY")
    # account = w3.eth.account.from_key(private_key)
    
    print("Deployment script ready. Add private key to execute.")

if __name__ == "__main__":
    deploy()
