requirements.txt
streamlit
web3
requests
pandas
python-dotenv
bcrypt
qrcode[pil]
watchdog
